Move ai code to entity class
